requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
frozenlist==1.3.3
github==1.2.7
idna==3.4
multidict==6.0.3
mysqlclient==2.1.1
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.4
urllib3==1.26.13
yarl==1.8.2
